+++ diPLSlib/models.py
@@ -8,6 +8,9 @@ diPLSlib model classes
 
 # Modules
 from sklearn.base import BaseEstimator, RegressorMixin
+from sklearn.utils.validation import check_array, check_X_y
+from sklearn.exceptions import NotFittedError
+from scipy.sparse import issparse, sparray
 import numpy as np
 import matplotlib.pyplot as plt
 from diPLSlib import functions as algo
@@ -29,9 +32,8 @@ class DIPLS(RegressorMixin, BaseEstimator):
     A : int
         Number of latent variables to be used in the model.
 
-    l : Union[int, List[int]], default=0
-        Regularization parameter. Either a single value or a list of different
-        values for each latent variable (LV).
+    l : float or tuple with len(l)=A, default=0
+        Regularization parameter. If a single value is provided, the same regularization is applied to all latent variables.
 
     centering : bool, default=True
             If True, source and target domain data are mean-centered.
@@ -54,16 +56,16 @@ class DIPLS(RegressorMixin, BaseEstimator):
     Attributes
     ----------
 
-    n : int
+    n_ : int
         Number of samples in `x`.
 
-    ns : int
+    ns_ : int
         Number of samples in `xs`.
 
-    nt : int
+    nt_ : int
         Number of samples in `xt`.
 
-    k : int
+    n_features_in_ : int
         Number of features (variables) in `x`.
 
     mu_ : ndarray of shape (n_features,)
@@ -123,6 +125,9 @@ class DIPLS(RegressorMixin, BaseEstimator):
     discrepancy_ : ndarray
         The variance discrepancy between source and target domain projections.
 
+    is_fitted_ : bool, default=False
+        Whether the model has been fitted to data.
+
 
     References
     ----------
@@ -142,15 +147,14 @@ class DIPLS(RegressorMixin, BaseEstimator):
     >>> xs = np.random.rand(100, 10)
     >>> xt = np.random.rand(50, 10)
     >>> X = np.random.rand(10, 10)
-    >>> model = DIPLS(A=5, l=[10])
+    >>> model = DIPLS(A=5, l=(10))
     >>> model.fit(x, y, xs, xt)
-    DIPLS(A=5, l=[10])
-
+    DIPLS(A=5, l=10)
     >>> xtest = np.array([5, 7, 4, 3, 2, 1, 6, 8, 9, 10]).reshape(1, -1)
     >>> yhat = model.predict(xtest)
     """
 
-    def __init__(self, A=2, l=[0], centering=True, heuristic=False, target_domain=0, rescale='Target'):
+    def __init__(self, A=2, l=0, centering=True, heuristic=False, target_domain=0, rescale='Target'):
         # Model parameters
         self.A = A
         self.l = l
@@ -158,9 +162,10 @@ class DIPLS(RegressorMixin, BaseEstimator):
         self.heuristic = heuristic
         self.target_domain = target_domain
         self.rescale = rescale
+        
 
 
-    def fit(self, X, y, xs, xt):
+    def fit(self, X, y, xs=None, xt=None):
         """
         Fit the DIPLS model.
 
@@ -177,11 +182,11 @@ class DIPLS(RegressorMixin, BaseEstimator):
             Response variable corresponding to the input data `x`.
 
         xs : ndarray of shape (n_samples_source, n_features)
-            Source domain X-data.
+            Source domain X-data. If not provided, defaults to `X`.
 
         xt : Union[ndarray of shape (n_samples_target, n_features), List[ndarray]]
             Target domain X-data. Can be a single target domain or a list of arrays 
-            representing multiple target domains.
+            representing multiple target domains. If not provided, defaults to `X`.
 
 
         Returns
@@ -189,53 +194,100 @@ class DIPLS(RegressorMixin, BaseEstimator):
         self : object
             Fitted model instance.
         """
-
-        # Preliminaries
-        self.n, self.k = X.shape
-        self.ns, _ = xs.shape        
         
-        self.x = X
-        self.y = y
-        self.xs = xs
-        self.xt = xt
-        self.b0_ = np.mean(self.y)
+        # Check for sparse input
+        if issparse(X):
+
+            raise ValueError("Sparse input is not supported. Please convert your data to dense format.")
+ 
+        # Validate input arrays
+        X, y = check_X_y(X, y, ensure_2d=True, allow_nd=False, accept_large_sparse=False, accept_sparse=False, force_all_finite=True)
+        
+
+        # Check if source and target data are provided
+        if xs is None:
+
+            xs = X
+
+        if xt is None:
+
+            xt = X
+
+        # Validate source and target arrays
+        xs = check_array(xs, ensure_2d=True, allow_nd=False, accept_large_sparse=False, accept_sparse=False, force_all_finite=True)
+        xs = np.atleast_2d(xs) if xs is not None else X
+        if isinstance(xt, list):
+            xt = [check_array(x, ensure_2d=True, allow_nd=False, accept_large_sparse=False, accept_sparse=False, force_all_finite=True) for x in xt]
+        else:
+            xt = check_array(xt, ensure_2d=True, allow_nd=False, accept_large_sparse=False, accept_sparse=False, force_all_finite=True)
+        xt = [np.atleast_2d(x) for x in xt] if isinstance(xt, list) else np.atleast_2d(xt) if xt is not None else X
+
+        # Flatten y to 1D array
+        y = np.ravel(y)
+
+        # Check for complex data
+        if np.iscomplexobj(X) or np.iscomplexobj(y) or np.iscomplexobj(xs) or np.iscomplexobj(xt):
+            
+            raise ValueError("Complex data not supported")
+        
+        
+        # Check if source and target data are provided
+        if xs is None:
+
+            xs = X
+
+        if xt is None:
+
+            xt = X
+        
+        
+        # Preliminaries
+        self.n_, self.n_features_in_ = X.shape
+        self.ns_, _ = xs.shape        
+        
+        self.x_ = X
+        self.y_ = y
+        self.xs_ = xs
+        self.xt_ = xt
+        self.b0_ = np.mean(self.y_)
 
         # Mean centering
         if self.centering:
 
-            self.mu_ = np.mean(self.x, axis=0)
-            self.mu_s_ = np.mean(self.xs, axis=0)
-            self.x = self.x - self.mu_
-            self.xs = self.xs - self.mu_s_
-            y = self.y - self.b0_
+            self.mu_ = np.mean(self.x_, axis=0)
+            self.mu_s_ = np.mean(self.xs_, axis=0)
+            self.x_ = self.x_ - self.mu_
+            self.xs_ = self.xs_ - self.mu_s_
+            y = self.y_ - self.b0_
 
             # Mutliple target domains
-            if isinstance(self.xt, list):
+            if isinstance(self.xt_, list):
                 
-                self.nt, _ = xt[0].shape
-                self.mu_t_ = [np.mean(x, axis=0) for x in self.xt]
-                self.xt = [x - mu for x, mu in zip(self.xt, self.mu_t_)]
+                self.nt_, _ = xt[0].shape
+                self.mu_t_ = [np.mean(x, axis=0) for x in self.xt_]
+                self.xt_ = [x - mu for x, mu in zip(self.xt_, self.mu_t_)]
             
             else:
 
-                self.nt, _ = xt.shape
-                self.mu_t_ = np.mean(self.xt, axis=0)
-                self.xt = self.xt - self.mu_t_
+                self.nt_, _ = xt.shape
+                self.mu_t_ = np.mean(self.xt_, axis=0)
+                self.xt_ = self.xt_ - self.mu_t_
 
         else:
 
-            y = self.y
+            y = self.y_
         
 
-        x = self.x 
-        xs = self.xs
-        xt = self.xt
+        x = self.x_ 
+        xs = self.xs_
+        xt = self.xt_
 
     
         # Fit model
-        results = algo.dipals(x, y, xs, xt, self.A, self.l, heuristic=self.heuristic, target_domain=self.target_domain)
+        results = algo.dipals(x, y.reshape(-1,1), xs, xt, self.A, self.l, heuristic=self.heuristic, target_domain=self.target_domain)
         self.b_, self.T_, self.Ts_, self.Tt_, self.W_, self.P_, self.Ps_, self.Pt_, self.E_, self.Es_, self.Et_, self.Ey_, self.C_, self.opt_l_, self.discrepancy_ = results
         
+        self.is_fitted_ = True        
         return self
 
             
@@ -259,13 +311,23 @@ class DIPLS(RegressorMixin, BaseEstimator):
             Predicted response values for the test data.
 
         """
+        if not hasattr(self, 'is_fitted_') or not self.is_fitted_:
+            raise NotFittedError("This DIPLS instance is not fitted yet. Call 'fit' with appropriate arguments before using this estimator.")
+        
+        
+        # Check for sparse input
+        if issparse(X):
+            raise ValueError("Sparse input is not supported. Please convert your data to dense format.")
+
+        # Validate input array
+        X = check_array(X, ensure_2d=True, allow_nd=False, force_all_finite=True)
         
         # Rescale Test data 
         if(type(self.rescale) is str):
 
             if(self.rescale == 'Target'):
 
-                if(type(self.xt) is list):
+                if(type(self.xt_) is list):
 
                     if(self.target_domain==0):
 
@@ -298,6 +360,8 @@ class DIPLS(RegressorMixin, BaseEstimator):
         
         yhat = Xtest@self.b_ + self.b0_
 
+        # Ensure the shape of yhat matches the shape of y
+        yhat = np.ravel(yhat)
 
         return yhat
 
@@ -314,10 +378,8 @@ class GCTPLS(DIPLS):
     Parameters
     ----------
 
-    l : Union[int, List[int]], default=0
-        Regularization parameter. Can be a single value or a list of different
-        values for each latent variable (LV). This parameter controls the degree
-        of regularization applied during the fitting process.
+     l : float or tuple with len(l)=A, default=0
+        Regularization parameter. If a single value is provided, the same regularization is applied to all latent variables.
 
     centering : bool, default=True
         If True, source and target domain data are mean-centered before fitting.
@@ -336,16 +398,16 @@ class GCTPLS(DIPLS):
     Attributes
     ----------
 
-    n : int
+    n_ : int
         Number of samples in `x`.
 
-    ns : int
+    ns_ : int
         Number of samples in `xs`.
 
-    nt : int
+    nt_ : int
         Number of samples in `xt`.
 
-    k : int
+    n_features_in_ : int
         Number of features (variables) in `x`.
 
     mu_ : ndarray of shape (n_features,)
@@ -405,6 +467,9 @@ class GCTPLS(DIPLS):
     discrepancy_ : ndarray
         The variance discrepancy between source and target domain projections.
 
+    is_fitted_ : bool, default=False
+        Whether the model has been fitted to data.
+
 
     References
     ----------
@@ -420,14 +485,14 @@ class GCTPLS(DIPLS):
     >>> y = np.random.rand(100, 1)
     >>> xs = np.random.rand(80, 10)
     >>> xt = np.random.rand(80, 10)
-    >>> model = GCTPLS(A=3, l=[10])
+    >>> model = GCTPLS(A=3, l=(2,5,7))
     >>> model.fit(x, y, xs, xt)
-    GCTPLS(A=3, l=[10])
+    GCTPLS(A=3, l=(2, 5, 7))
     >>> xtest = np.array([5, 7, 4, 3, 2, 1, 6, 8, 9, 10]).reshape(1, -1)
     >>> yhat = model.predict(xtest)
     """
 
-    def __init__(self, A=2, l=[0], centering=True, heuristic=False, rescale='Target'):
+    def __init__(self, A=2, l=0, centering=True, heuristic=False, rescale='Target'):
         # Model parameters
         self.A = A
         self.l = l
@@ -436,7 +501,7 @@ class GCTPLS(DIPLS):
         self.rescale = rescale
 
         
-    def fit(self, X, y, xs, xt):
+    def fit(self, X, y, xs=None, xt=None):
         """
         Fit the GCT-PLS model to data.
 
@@ -450,10 +515,10 @@ class GCTPLS(DIPLS):
             Response variable corresponding to the input data `x`.
 
         xs : ndarray of shape (n_sample_pairs, n_features)
-            Source domain X-data.
+            Source domain X-data. If not provided, defaults to `X`.
 
         xt : ndarray of shape (n_sample_pairs, n_features)
-            Target domain X-data. 
+            Target domain X-data. If not provided, defaults to `X`.
  
 
         Returns
@@ -462,41 +527,86 @@ class GCTPLS(DIPLS):
         self : object
             Fitted model instance.
         """
-        
-        # Preliminaries
-        self.n, self.k = X.shape
-        self.ns, _ = xs.shape
-        self.nt, _ = xt.shape
+        # Check for sparse input
+        if issparse(X):
 
-        if self.ns != self.nt:
+            raise ValueError("Sparse input is not supported. Please convert your data to dense format.")
+
+        # Validate input arrays
+        X, y = check_X_y(X, y, ensure_2d=True, allow_nd=False, accept_large_sparse=False, accept_sparse=False, force_all_finite=True)
+        
+        # Check if source and target data are provided
+        if xs is None:
+
+            xs = X
+
+        if xt is None:
+
+            xt = X
+
+        # Validate source and target arrays
+        xs = check_array(xs, ensure_2d=True, allow_nd=False, accept_large_sparse=False, accept_sparse=False, force_all_finite=True)
+        xs = np.atleast_2d(xs) if xs is not None else X
+        if isinstance(xt, list):
+            xt = [check_array(x, ensure_2d=True, allow_nd=False, accept_large_sparse=False, accept_sparse=False, force_all_finite=True) for x in xt]
+        else:
+            xt = check_array(xt, ensure_2d=True, allow_nd=False, accept_large_sparse=False, accept_sparse=False, force_all_finite=True)
+        xt = [np.atleast_2d(x) for x in xt] if isinstance(xt, list) else np.atleast_2d(xt) if xt is not None else X
+
+        # Flatten y to 1D array
+        y = np.ravel(y)
+
+        # Check for complex data
+        if np.iscomplexobj(X) or np.iscomplexobj(y) or np.iscomplexobj(xs) or np.iscomplexobj(xt):
+            
+            raise ValueError("Complex data not supported")
+        
+        
+        # Check if source and target data are provided
+        if xs is None:
+
+            xs = X
+
+        if xt is None:
+
+            xt = X
+        
+
+        # Preliminaries
+        self.n_, self.n_features_in_ = X.shape
+        self.ns_, _ = xs.shape        
+        self.nt_, _ = xt.shape
+
+        if self.ns_ != self.nt_:
             raise ValueError("The number of samples in the source domain (ns) must be equal to the number of samples in the target domain (nt).")
         
-        self.x = X
-        self.y = y
-        self.xs = xs
-        self.xt = xt
-        self.b0_ = np.mean(self.y)
-        self.mu_ = np.mean(self.x, axis=0)
-        self.mu_s_ = np.mean(self.x, axis=0)
-        self.mu_t_ = np.mean(self.x, axis=0)
+        self.x_ = X
+        self.y_ = y
+        self.xs_ = xs
+        self.xt_ = xt
+        self.b0_ = np.mean(self.y_)
+        self.mu_ = np.mean(self.x_, axis=0)
+        self.mu_s_ = np.mean(self.xs_, axis=0)
+        self.mu_t_ = np.mean(self.xt_, axis=0)
 
         # Mean Centering
         if self.centering is True:
             
-            x = self.x[...,:] - self.mu_
-            y = self.y - self.b0_
+            x = self.x_[...,:] - self.mu_
+            y = self.y_ - self.b0_
 
         else: 
             
-            x = self.x
-            y = self.y
+            x = self.x_
+            y = self.y_
 
-        xs = self.xs
-        xt = self.xt
+        xs = self.xs_
+        xt = self.xt_
             
         # Fit model and store matrices
-        results = algo.dipals(x, y, xs, xt, self.A, self.l, heuristic=self.heuristic, laplacian=True)
+        results = algo.dipals(x, y.reshape(-1,1), xs, xt, self.A, self.l, heuristic=self.heuristic, laplacian=True)
         self.b_, self.T_, self.Ts_, self.Tt_, self.W_, self.P_, self.Ps_, self.Pt_, self.E_, self.Es_, self.Et_, self.Ey_, self.C_, self.opt_l_, self.discrepancy_ = results
 
+        self.is_fitted_ = True  # Set the is_fitted attribute to True
         return self