+++ tests/test_doctests.py
@@ -3,11 +3,15 @@ import unittest
 import diPLSlib.models
 import diPLSlib.functions
 import diPLSlib.utils.misc
+from sklearn.utils.estimator_checks import check_estimator
+from diPLSlib.models import DIPLS, GCTPLS
 import nbformat
 from nbconvert.preprocessors import ExecutePreprocessor
 import os
 
 class TestDocstrings(unittest.TestCase):
+
+    # Test if all docstring examples run without errors
     def test_docstrings(self):
 
         # Run doctests across all modules in your_package
@@ -15,6 +19,7 @@ class TestDocstrings(unittest.TestCase):
         doctest.testmod(diPLSlib.functions)
         doctest.testmod(diPLSlib.utils.misc)
 
+    # Test if all notebooks run without errors
     def test_notebooks(self):
         # List of notebooks to test
         notebooks = [
@@ -33,6 +38,12 @@ class TestDocstrings(unittest.TestCase):
                 os.chdir(root_dir)
 
                 ep.preprocess(nb, {'metadata': {'path': './notebooks/'}})
+
+        # Test if diPLSlib.model classes pass check_estimator
+        def test_check_estimator(self):
+
+            for model in [DIPLS, GCTPLS]:
+                check_estimator(model)
         
 
 if __name__ == '__main__':